Extend GitRepoTests to allow include mode
The GitRepoTests can be initialized with a ValidateWorkingTree
mode that is currently a boolean. If enabled, we will add extra
checks to make sure the working tree of our virtual enlistment
matches the enlistment from a vanilla Git repo.

To extend this to the "include" mode, we need to first modify
this boolean into an enum and add it as an extra option.

Second, we need to be very careful about what we are comparing.
We will no longer virtualize outside of the included folders.
Be careful to include files in the root directory.

The functional tests can then be run with "--partial-mode" to
exercise only this option, which will isolate the test failures
to this setting AND keep the test suite from getting too slow.

Currently, the partial mode space is probably wider than
necessary and can be reduced in later iterations. This
does provide _some_ coverage.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>