GVFSVerb: Use OAuth credentials by default
To more rapidly adopt OAuth tokens, set that as the default for GVFS
repos. This will update on mount, so all users will update to this mode
when they upgrade to a version including this commit.

This may cause some initial frustration as the first time I ran a fetch
in OAuth mode my local clone had a failure with GCM and defaulted to
username/password checks over command line. The second fetch worked just
fine, though.